Add shading adjustment to clearsky calculations

+++ shared/lib_pv_io_manager.h
@@ -283,7 +283,7 @@ struct PVSystem_IO
 
 	size_t numberOfSubarrays;
 	size_t numberOfInverters;
-    size_t numberOfInvertersClipping;
+    double numberOfInvertersClipping;
     double dcNameplate;
 
 	Irradiance_IO * Irradiance;
